refactor date-parse.js
